chore: enums in LightTable
