+++ packages/frontend/src/components/common/LightTable/types.ts
@@ -1,0 +1,10 @@
+export enum SectionType {
+    Head = 'head',
+    Body = 'body',
+    Footer = 'footer',
+}
+
+export enum CellType {
+    Head = 'head',
+    Data = 'data',
+}
